+++ src/VueDatePicker/constants/defaults.ts
@@ -1,4 +1,4 @@
-import { WeekStart } from '@/types';
+import { WeekStart } from '.';
 import { enUS } from 'date-fns/locale';
 
 export const defaultRangeOptions = {
@@ -149,7 +149,7 @@ export const basePropDefaults = {
 };
 
 export const propDefaults = {
-    weekStart: WeekStart.Sunday,
+    weekStart: WeekStart.Monday,
     yearRange: (): [number, number] => [1900, 2100],
     ui: () => ({}),
     locale: () => enUS,
