+++ src/VueDatePicker/constants/index.ts
@@ -53,3 +53,13 @@ export enum MAP_KEY_FORMAT {
     YEAR = 'yyyy',
     DATE = 'dd-MM-yyyy',
 }
+
+export enum WeekStart {
+    Sunday = 0,
+    Monday = 1,
+    Tuesday = 2,
+    Wednesday = 3,
+    Thursday = 4,
+    Friday = 5,
+    Saturday = 6,
+}
