--- src/VueDatePicker/types/enums.ts
@@ -1,9 +1,0 @@
-export enum WeekStart {
-    Sunday = 0,
-    Monday = 1,
-    Tuesday = 2,
-    Wednesday = 3,
-    Thursday = 4,
-    Friday = 5,
-    Saturday = 6,
-}
