+++ src/VueDatePicker/types/index.ts
@@ -4,4 +4,3 @@ export * from './configs.ts';
 export * from './generic.ts';
 export * from './picker.ts';
 export * from './emits.ts';
-export * from './enums.ts';
