+++ src/VueDatePicker/types/props.ts
@@ -21,7 +21,7 @@ import type {
 import type { CalendarWeek, Marker, TimeModel, PresetDate, HighlightFn, DisabledTimesFn } from '@/types/picker.ts';
 import type { PickerSection, DateValue, ModelValue, SixWeekMode } from '@/types/generic.ts';
 import { basePropDefaults, type propDefaults } from '@/constants/defaults.ts';
-import type { WeekStart } from '@/types/enums.ts';
+import type { WeekStart } from '@/constants';
 
 export interface RootProps {
     multiCalendars?: boolean | number | string | Partial<MultiCalendarsConfig>;
