+++ src/index.ts
@@ -1,3 +1,4 @@
 export { default as VueDatePicker } from './VueDatePicker/VueDatePickerRoot.vue';
 export { TZDate } from '@date-fns/tz';
 export * from './VueDatePicker/types';
+export { WeekStart } from './VueDatePicker/constants';
